Enhance code coverage, appease SonarCloud
